src/requirements.txt
torch
sentence-transformers[openvino]
python-dotenv
pandas
nltk
jsonlines
streamlit
chromadb